requirements.txt
fastapi
uvicorn
pydantic
pageindex
python-dotenv
pydantic-settings
google-genai